Course crud complete

+++ app/Http/Controllers/CourseController.php
@@ -92,6 +92,10 @@ class CourseController extends Controller
     {
         $filters = $request->except('_token', '_method');
 
+        if ($course->subCourses()->exists()) {
+            return redirect()->route('courses.index', $filters)
+                ->with('error', 'Course can not be deleted as it has one or more subcourses.');
+        }
         $course->delete();
 
         return redirect()->route('courses.index', $filters)
